// src/descriptions/RuleDescription.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type RuleDescription =
  | {
      description?: string;
      constrains?:
        | {
            entity: {
              [k: string]: string;
            };
          }
        | {
            agent: {
              [k: string]: string;
            };
          }
        | {
            location: {
              [k: string]: string;
            };
          };
      constraints?: ExprDescription[];
    }
  | {
      description?: string;
      when:
        | {
            /**
             * @minItems 1
             */
            all: [ExprDescription, ...ExprDescription[]];
            [k: string]: unknown;
          }
        | {
            /**
             * @minItems 1
             */
            any: [ExprDescription, ...ExprDescription[]];
            [k: string]: unknown;
          }
        | ExprDescription;
      then: [ExprDescription, ...ExprDescription[]] | ExprDescription;
    };

export interface ExprDescription {
  expr: string;
  explain?: string;
  [k: string]: unknown;
}